+++ xstampp.astpa/src/xstampp/astpa/ui/causalfactors/CausalFactorsView.java
@@ -1,8 +1,8 @@
 /*******************************************************************************
  * 
  * Copyright (c) 2013-2017 A-STPA Stupro Team Uni Stuttgart (Lukas Balzer, Adam Grahovac, Jarkko
- * Heidenwag, Benedikt Markt, Jaqueline Patzek, Sebastian Sieber, Fabian Toth, Patrick
- * Wickenhäuser, Aliaksei Babkovich, Aleksander Zotov).
+ * Heidenwag, Benedikt Markt, Jaqueline Patzek, Sebastian Sieber, Fabian Toth, Patrick Wickenhäuser,
+ * Aliaksei Babkovich, Aleksander Zotov).
  * 
  * All rights reserved. This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v1.0 which accompanies this distribution, and is available at
@@ -48,7 +48,7 @@ import xstampp.usermanagement.api.AccessRights;
 
 /**
  * The view to add causal factors to control structure components, edit them and add links to the
- * related hazards.
+ * related hazards. This <b>extends</b> {@link CommonGridView}.
  * 
  * @author Benedikt Markt, Patrick Wickenhaeuser, Lukas Balzer
  */
@@ -91,8 +91,9 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
 
   @Override
   public void createPartControl(Composite parent) {
-    ICausalFactorDataModel dataModel = (ICausalFactorDataModel) ProjectManager.getContainerInstance()
-        .getDataModel(this.getProjectID());
+    ICausalFactorDataModel dataModel = (ICausalFactorDataModel) ProjectManager
+        .getContainerInstance().getDataModel(this.getProjectID());
+
     if (dataModel.getCausalFactorController().analyseFactorsPerUCA()) {
       setting = new UCAScenarioSetting();
     } else {
@@ -135,7 +136,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
   @Override
   protected String[] getCategoryArray() {
     return new String[] { "ALL", ComponentType.ACTUATOR.name(), ComponentType.CONTROLLER.name(),
-        ComponentType.CONTROLLED_PROCESS.name(), ComponentType.SENSOR.name(), CAUSALFACTORS, "UCA" };
+        ComponentType.CONTROLLED_PROCESS.name(), ComponentType.SENSOR.name(), CAUSALFACTORS,
+        "UCA" };
   }
 
   /**
@@ -164,6 +166,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
 
   @Override
   protected void fillTable() {
+    // The column labels are set depending on whether the user wants to use Scenarios, than an extra
+    // column for scenarios is added or not
     if (this.getDataModel().isUseScenarios()) {
       this.getGridWrapper().setColumnLabels(setting.getScenarioColumns());
     } else {
@@ -176,35 +180,25 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
         GridCellText cell = new GridCellText(component.getText());
         cell.setToolTip(component.getText());
         componentRow.addCell(0, cell);
-        if (createPrimaryFactorRows(component, componentRow)) {
+        // here all child rows containing the causal factors for the current component are created
+        if (setting.addPrimaryFactor(component, componentRow)) {
           getGridWrapper().addRow(componentRow);
         }
       }
     }
   }
 
-  /**
-   * 
-   * @param component
-   *          The {@link IRectangleComponent} for which the Causal Factor should be derived.
-   * @param componentRow
-   *          The parent row that has been added for the Component
-   * @param ucaCfLink_Component_ToCFmap
-   */
-  private boolean createPrimaryFactorRows(IRectangleComponent component, GridRow componentRow) {
-    return setting.addPrimaryFactor(component, componentRow);
-  }
-
   private void createScenarioRow(Link factorComponentLink, GridRow entryRow) {
     GridRow scenarioRow = new GridRow(this.getGridWrapper().getColumnLabels().length);
     GridCellText scenarioCell = new GridCellText("Add a new scenario");
-    Link ucaCFLink = getDataModel().getLinkController().getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK,
-        factorComponentLink.getLinkA());
+    Link ucaCFLink = getDataModel().getLinkController()
+        .getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK, factorComponentLink.getLinkA());
     IUnsafeControlAction uca = getDataModel().getControlActionController()
         .getUnsafeControlAction(ucaCFLink.getLinkA());
     scenarioCell.addCellButton(new CellButtonAddScenario(getDataModel(), factorComponentLink, uca));
     scenarioCell.addCellButton(new CellButtonLinking<ContentProviderScenarios>(getGridWrapper(),
-        new ContentProviderScenarios(getDataModel(), factorComponentLink, uca), factorComponentLink.getLinkA()));
+        new ContentProviderScenarios(getDataModel(), factorComponentLink, uca),
+        factorComponentLink.getLinkA()));
     scenarioRow.setColumnSpan(4, 1);
     scenarioRow.addCell(4, scenarioCell);
     entryRow.addChildRow(scenarioRow);
@@ -214,13 +208,14 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
    * 
    * @param entryRow
    *          the row in which the cells will be added
-   * @param ucaHazLink
+   * @param entryHazLink
    *          a Link of type {@link LinkingType#CausalEntryLink_HAZ_LINK}
    * @return
    */
-  private GridRow createHazardRow(GridRow entryRow, Link ucaHazLink) {
-    ITableModel hazard = getDataModel().getHazard(ucaHazLink.getLinkB());
-    String hazText = hazard != null ? hazard.getIdString() + " - " + hazard.getTitle() : "no hazard";
+  private GridRow createHazardRow(GridRow entryRow, Link entryHazLink) {
+    ITableModel hazard = getDataModel().getHazard(entryHazLink.getLinkB());
+    String hazText = hazard != null ? hazard.getIdString() + " - " + hazard.getTitle()
+        : "no hazard";
     GridCellText hazCell = new GridCellText(hazText);
     if (hazard != null && hazard.getDescription() != null) {
       String toolTip = hazard.getDescription();
@@ -237,12 +232,12 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
     }
     entryRow.addCell(3, hazCell);
 
-    CellEditorSingleSafetyConstraint cell = new CellEditorSingleSafetyConstraint(getGridWrapper(),
-        getDataModel(), ucaHazLink);
-    Link factorComponentLink = getDataModel().getLinkController().getLinkObjectFor(LinkingType.UcaCfLink_Component_LINK,
-        ucaHazLink.getLinkA());
-    Link ucaCfLink = getDataModel().getLinkController().getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK,
-        factorComponentLink.getLinkA());
+    CellEditorSafetyConstraint cell = new CellEditorSafetyConstraint(getGridWrapper(),
+        getDataModel(), entryHazLink);
+    Link factorComponentLink = getDataModel().getLinkController()
+        .getLinkObjectFor(LinkingType.UcaCfLink_Component_LINK, entryHazLink.getLinkA());
+    Link ucaCfLink = getDataModel().getLinkController()
+        .getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK, factorComponentLink.getLinkA());
     ITableModel actionForUca = getDataModel().getControlActionForUca(ucaCfLink.getLinkA());
     if (actionForUca != null && !checkAccess(actionForUca.getId(), AccessRights.WRITE)) {
       cell.setReadOnly(true);
@@ -251,7 +246,7 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
     entryRow.addCell(4, cell);
 
     Optional<Link> safetyOption = getDataModel().getLinkController()
-        .getRawLinksFor(LinkingType.CausalHazLink_SC2_LINK, ucaHazLink.getId()).stream()
+        .getRawLinksFor(LinkingType.CausalHazLink_SC2_LINK, entryHazLink.getId()).stream()
         .findFirst();
 
     IGridCell hintCell = new GridCellText("");
@@ -262,7 +257,7 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
     // in column 5 the Design Hint cell for the CausalHazLink_SC2_LINK is added
     entryRow.addCell(5, hintCell);
     // In column 6 the Note/Rational cell for the CausalEntryLink_HAZ_LINK is created
-    entryRow.addCell(6, new CellEditorFactorNote(getGridWrapper(), getDataModel(), ucaHazLink));
+    entryRow.addCell(6, new CellEditorFactorNote(getGridWrapper(), getDataModel(), entryHazLink));
     return null;
 
   }
@@ -278,11 +273,11 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
    * @return
    */
   private GridRow createSingleConstraintRow(GridRow entryRow, Link factorConstraintLink) {
-    Link factorComponentLink = getDataModel().getLinkController().getLinkObjectFor(LinkingType.UcaCfLink_Component_LINK,
-        factorConstraintLink.getLinkA());
+    Link factorComponentLink = getDataModel().getLinkController()
+        .getLinkObjectFor(LinkingType.UcaCfLink_Component_LINK, factorConstraintLink.getLinkA());
     addHazardCell(entryRow, factorComponentLink);
-    Link ucaCfLink = getDataModel().getLinkController().getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK,
-        factorComponentLink.getLinkA());
+    Link ucaCfLink = getDataModel().getLinkController()
+        .getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK, factorComponentLink.getLinkA());
     CellEditorSingleSafetyConstraint cell = new CellEditorSingleSafetyConstraint(getGridWrapper(),
         getDataModel(), factorConstraintLink);
     ITableModel actionForUca = getDataModel().getControlActionForUca(ucaCfLink.getLinkA());
@@ -292,7 +287,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
     }
     entryRow.addCell(4, cell);
 
-    CellEditorFactorNote hintCell = new CellEditorFactorNote(getGridWrapper(), getDataModel(), factorConstraintLink);
+    CellEditorFactorNote hintCell = new CellEditorFactorNote(getGridWrapper(), getDataModel(),
+        factorConstraintLink);
     if (factorConstraintLink.getLinkB() == null) {
       hintCell.setReadOnly(true);
     } else {
@@ -306,8 +302,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
   }
 
   private void addHazardCell(GridRow entryRow, Link factorComponentLink) {
-    Link ucaCfLink = getDataModel().getLinkController().getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK,
-        factorComponentLink.getLinkA());
+    Link ucaCfLink = getDataModel().getLinkController()
+        .getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK, factorComponentLink.getLinkA());
     String hazString = "";
     String tooltip = "";
     for (UUID hazId : getDataModel().getLinkController().getLinksFor(LinkingType.UCA_HAZ_LINK,
@@ -317,9 +313,9 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
       hazString += hazard.getIdString();
       tooltip += tooltip.isEmpty() ? "" : "\n";
       tooltip += hazard.getIdString();
-      for(int i=0; i < hazard.getTitle().length();) {
-        int nextI = hazard.getTitle().indexOf(' ', i+ 30);
-        nextI = nextI == -1 ? hazard.getTitle().length(): nextI;
+      for (int i = 0; i < hazard.getTitle().length();) {
+        int nextI = hazard.getTitle().indexOf(' ', i + 30);
+        nextI = nextI == -1 ? hazard.getTitle().length() : nextI;
         tooltip += "\n   " + hazard.getTitle().substring(i, nextI);
         i = nextI;
       }
@@ -339,8 +335,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
    * @return
    */
   private void createScenarioRow(GridRow entryRow, Link scenarioLink) {
-    Link factorComponentLink = getDataModel().getLinkController().getLinkObjectFor(LinkingType.UcaCfLink_Component_LINK,
-        scenarioLink.getLinkA());
+    Link factorComponentLink = getDataModel().getLinkController()
+        .getLinkObjectFor(LinkingType.UcaCfLink_Component_LINK, scenarioLink.getLinkA());
     addHazardCell(entryRow, factorComponentLink);
     ScenarioType type = getDataModel().getExtendedDataController()
         .getScenarioType(scenarioLink.getLinkB());
@@ -375,8 +371,14 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
     }
   }
 
+  /**
+   * This setting object decides whether the causal Factors are displayed like:<br> <ul> <li>1 CF
+   * contains n UCA, or <li>1 UCA contains n CF </ul>
+   * 
+   */
   interface CausalMethodologySetting {
     /**
+     * This s
      * 
      * @param component
      *          The {@link IRectangleComponent} for which the Causal Factor should be derived.
@@ -408,6 +410,9 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
       return _withoutScenarioColumns.toArray(new String[0]);
     }
 
+    /**
+     * This implementation of the method creates a
+     */
     @Override
     public boolean addPrimaryFactor(IRectangleComponent component, GridRow componentRow) {
       Map<ITableModel, List<Link>> ucaCfLink_Component_To_CF_map = getDataModel()
@@ -416,7 +421,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
       for (ITableModel factor : ucaCfLink_Component_To_CF_map.keySet()) {
         if (factor != null && !isFiltered(factor.getText(), CAUSALFACTORS)) {
           GridRow causalFactorRow;
-          causalFactorRow = new GridRow(getGridWrapper().getColumnLabels().length, 1, new int[] { 1 });
+          causalFactorRow = new GridRow(getGridWrapper().getColumnLabels().length, 1,
+              new int[] { 1 });
           CellEditorCausalFactor cell = new CellEditorCausalFactor(getGridWrapper(), getDataModel(),
               factor.getId());
           // If the current user of the project is an Administrator than he/she can delete or edit
@@ -431,11 +437,10 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
           for (Link factorComponentLink : factorComponentLinkList) {
             /*
              * Depending on whether the entry is linked to a uca or not the uca column is filled and
-             * the
-             * hazards are either based on the uca or linkable
+             * the hazards are either based on the uca or linkable
              */
-            Link ucaCFLink = getDataModel().getLinkController()
-                .getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK, factorComponentLink.getLinkA());
+            Link ucaCFLink = getDataModel().getLinkController().getLinkObjectFor(
+                LinkingType.UCA_CausalFactor_LINK, factorComponentLink.getLinkA());
             IUnsafeControlAction uca = getDataModel().getControlActionController()
                 .getUnsafeControlAction(ucaCFLink.getLinkA());
             ucaList.add(uca);
@@ -470,7 +475,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
      * @param factorComponentLink
      *          a Link of type {@link LinkingType#UcaCfLink_Component_LINK}
      */
-    public void addSecondaryFactor(GridRow factorRow, Link factorComponentLink, IUnsafeControlAction uca) {
+    public void addSecondaryFactor(GridRow factorRow, Link factorComponentLink,
+        IUnsafeControlAction uca) {
 
       GridRow entryRow = new GridRow(getGridWrapper().getColumnLabels().length, 1, new int[] { 2 });
       CellEditorCausalEntry cell = new CellEditorCausalEntry(getGridWrapper(), getDataModel(),
@@ -482,7 +488,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
 
       entryRow.addCell(2, cell);
 
-      LinkingType entryType = getDataModel().isUseScenarios() ? LinkingType.CausalEntryLink_Scenario_LINK
+      LinkingType entryType = getDataModel().isUseScenarios()
+          ? LinkingType.CausalEntryLink_Scenario_LINK
           : LinkingType.CausalEntryLink_HAZ_LINK;
       // If a CausalEntryLink_SC2_LINK exists for the given UcaCfLink_Component_LINK than and
       // scenarios are not used than a single row containing only that safety constraint is created
@@ -506,8 +513,10 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
         GridRow row = new GridRow(getGridWrapper().getColumnLabels().length, 1);
         switch (entryType) {
         case CausalEntryLink_Scenario_LINK: {
-          if (getDataModel().getExtendedDataController().getRefinedScenario(link.getLinkB()) == null) {
-            ProjectManager.getLOGGER().error("Causal Factor Scenario link with illegal Scenario id!");
+          if (getDataModel().getExtendedDataController()
+              .getRefinedScenario(link.getLinkB()) == null) {
+            ProjectManager.getLOGGER()
+                .error("Causal Factor Scenario link with illegal Scenario id!");
           } else {
             createScenarioRow(row, link);
           }
@@ -568,11 +577,15 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
     public boolean addPrimaryFactor(IRectangleComponent component, GridRow componentRow) {
       boolean addComponentRow = false;
       Map<IUnsafeControlAction, List<Link>> ucaCfLink_Component_To_UCA_map = getDataModel()
-          .getCausalFactorController()
-          .getUCABasedMap(component, getDataModel().getLinkController(), getDataModel().getControlActionController());
+          .getCausalFactorController().getUCABasedMap(component, getDataModel().getLinkController(),
+              getDataModel().getControlActionController());
       List<IUnsafeControlAction> ucaList = new ArrayList<>();
+      // iterates all UCAs stored in the data model and included in the
+      // ucaCfLink_Component_To_UCA_map as keys to a list of links to causal factors
       for (IUnsafeControlAction uca : ucaCfLink_Component_To_UCA_map.keySet()) {
         if (uca != null && !isFiltered(uca.getId(), "UCA")) {
+          // addComponentRow is only set true if the uca is not filtered so it wont appear in the
+          // grid
           addComponentRow = true;
           List<Link> causalEntryLinks = ucaCfLink_Component_To_UCA_map.get(uca);
           ucaList.add(uca);
@@ -582,9 +595,9 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
           CellEditorCausalEntry cell = new CellEditorCausalEntry(getGridWrapper(), getDataModel(),
               causalEntryLinks, uca, uca.getId());
           String tooltip = uca.getIdString() + "\n\n" + ca.getIdString();
-          for(int i=0; i < ca.getTitle().length();) {
-            int nextI = ca.getTitle().indexOf(' ', i+ 200);
-            nextI = nextI == -1 ? ca.getTitle().length(): nextI;
+          for (int i = 0; i < ca.getTitle().length();) {
+            int nextI = ca.getTitle().indexOf(' ', i + 200);
+            nextI = nextI == -1 ? ca.getTitle().length() : nextI;
             tooltip += "\n   " + ca.getTitle().substring(i, nextI);
             i = nextI;
           }
@@ -600,8 +613,7 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
           for (Link link : causalEntryLinks) {
             /*
              * Depending on whether the entry is linked to a uca or not the uca column is filled and
-             * the
-             * hazards are either based on the uca or linkable
+             * the hazards are either based on the uca or linkable
              */
             Link ucaCFLink = getDataModel().getLinkController()
                 .getLinkObjectFor(LinkingType.UCA_CausalFactor_LINK, link.getLinkA());
@@ -615,7 +627,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
           // A new row is added to the factorRow for adding additional entries
           GridRow addEntriesRow = new GridRow(getGridWrapper().getColumnLabels().length);
           if (checkAccess(AccessRights.WRITE)) {
-            addEntriesRow.addCell(2, new GridCellButtonAddCausalFactor(component, getDataModel(), uca.getId()));
+            addEntriesRow.addCell(2,
+                new GridCellButtonAddCausalFactor(component, getDataModel(), uca.getId()));
           }
 
           addEntriesRow.setColumnSpan(2, getGridWrapper().getColumnLabels().length - 3);
@@ -627,8 +640,8 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
       // the filter is not set to filter for a specific UCA
       if (!getActiveCategory().equals("UCA")) {
         GridRow buttonRow = new GridRow(getGridWrapper().getColumnLabels().length);
-        buttonRow.addCell(1, new GridCellButtonAddUCAEntry(component, null,
-            getDataModel(), getGrid(), ucaList));
+        buttonRow.addCell(1,
+            new GridCellButtonAddUCAEntry(component, null, getDataModel(), getGrid(), ucaList));
         buttonRow.setColumnSpan(1, getGridWrapper().getColumnLabels().length - 2);
         componentRow.addChildRow(buttonRow);
         addComponentRow = true;
@@ -647,9 +660,11 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
      */
     public void addSecondaryFactor(GridRow factorRow, Link factorComponentLink, Link ucaCFLink,
         ITableModel factor) {
-      //if the link between uca and causal Factor contains a non existent uuid than it should be deleted,
-      //however if the value is null than the link is treated as the initial link that links the uca to the
-      //Component without linking it to a causal factor 
+      // if the link between uca and causal Factor contains a non existent uuid than it should be
+      // deleted,
+      // however if the value is null than the link is treated as the initial link that links the
+      // uca to the
+      // Component without linking it to a causal factor
       if (ucaCFLink.getLinkB() != null && getDataModel().getCausalFactorController()
           .getCausalFactor(ucaCFLink.getLinkB()) instanceof BadReferenceModel) {
         getDataModel().getLinkController().deleteLink(factorComponentLink.getLinkType(),
@@ -657,16 +672,17 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
         getDataModel().getLinkController().deleteLink(ucaCFLink.getLinkType(), ucaCFLink.getId());
         return;
       }
-      if(ucaCFLink.getLinkB() == null) {
+      if (ucaCFLink.getLinkB() == null) {
         return;
       }
       GridRow entryRow = new GridRow(getGridWrapper().getColumnLabels().length, 1, new int[] { 2 });
-      CellEditorCausalFactor cell = new CellEditorCausalFactor(getGridWrapper(), getDataModel(), factor.getId(),
-          factorComponentLink);
+      CellEditorCausalFactor cell = new CellEditorCausalFactor(getGridWrapper(), getDataModel(),
+          factor.getId(), factorComponentLink);
 
       entryRow.addCell(2, cell);
 
-      LinkingType entryType = getDataModel().isUseScenarios() ? LinkingType.CausalEntryLink_Scenario_LINK
+      LinkingType entryType = getDataModel().isUseScenarios()
+          ? LinkingType.CausalEntryLink_Scenario_LINK
           : LinkingType.CausalEntryLink_HAZ_LINK;
       // If a CausalEntryLink_SC2_LINK exists for the given UcaCfLink_Component_LINK than and
       // scenarios are not used than a single row containing only that safety constraint is created
@@ -682,8 +698,10 @@ public class CausalFactorsView extends CommonGridView<ICausalFactorDataModel> {
         GridRow row = new GridRow(getGridWrapper().getColumnLabels().length, 1);
         switch (entryType) {
         case CausalEntryLink_Scenario_LINK: {
-          if (getDataModel().getExtendedDataController().getRefinedScenario(link.getLinkB()) == null) {
-            ProjectManager.getLOGGER().error("Causal Factor Scenario link with illegal Scenario id!");
+          if (getDataModel().getExtendedDataController()
+              .getRefinedScenario(link.getLinkB()) == null) {
+            ProjectManager.getLOGGER()
+                .error("Causal Factor Scenario link with illegal Scenario id!");
           } else {
             createScenarioRow(row, link);
           }